+++ src/Resources/app/administration/src/module/sw-cms/blocks/ultimate-cms-tools/related-products/preview/index.js
@@ -1,0 +1,6 @@
+import template from './sw-cms-preview-related-products.html.twig';
+import './sw-cms-preview-related-products.scss';
+
+Shopware.Component.register('sw-cms-preview-related-products', {
+    template
+});